\name{starma-internal}
\alias{stacfCPP}
\alias{starmaCPP}
\alias{stcovCPP}
\alias{stpacfCPP}
\title{Internal \code{\link{starma-package}} functions}
\description{
Internal \code{\link{starma-package}} functions
} 
\details{
These functions are not user-level and should not be directly called.
}
\keyword{starma}
\keyword{internal}